Add more time for container shutdown
