+++ src/main/java/de/rngcntr/janusbench/backend/configuration/ComposeConfiguration.java
@@ -3,6 +3,8 @@ package de.rngcntr.janusbench.backend.configuration;
 import de.rngcntr.janusbench.backend.Index;
 import de.rngcntr.janusbench.backend.Storage;
 import java.io.File;
+import java.time.Duration;
+
 import org.testcontainers.containers.ContainerLaunchException;
 import org.testcontainers.containers.DockerComposeContainer;
 import org.testcontainers.containers.wait.strategy.Wait;
@@ -48,7 +50,9 @@ public class ComposeConfiguration extends Configuration {
 
         environment =
             getEnvironment()
-                .withExposedService("janusgraph", JANUSGRAPH_PORT)
+                .withExposedService(
+                    "janusgraph", JANUSGRAPH_PORT,
+                    Wait.forListeningPort().withStartupTimeout(Duration.ofMinutes(5)))
                 .waitingFor("janusgraph", Wait.forLogMessage(".*Channel started at port.*", 1))
                 .withLocalCompose(true);
